+++ solvers/CompressibleFlowSolver/Preconditioner/PreconCfs.cpp
@@ -39,6 +39,12 @@ using namespace std;
 
 namespace Nektar
 {
+PreconCfsFactory &GetPreconCfsFactory()
+{
+    static PreconCfsFactory instance;
+    return instance;
+}
+
 PreconCfs::PreconCfs(const Array<OneD, MultiRegions::ExpListSharedPtr> &pFields,
                      const LibUtilities::SessionReaderSharedPtr &pSession,
                      const LibUtilities::CommSharedPtr &vComm)
@@ -47,56 +53,6 @@ PreconCfs::PreconCfs(const Array<OneD, MultiRegions::ExpListSharedPtr> &pFields,
     m_verbose  = pSession->DefinesCmdLineArgument("verbose");
     m_spacedim = pFields[0]->GetGraph()->GetSpaceDimension();
     pSession->LoadParameter("PreconMatFreezNumb", m_PreconMatFreezNumb, 200);
-}
-void PreconCfs::v_InitObject()
-{
-}
-
-void PreconCfs::DoNullPrecon(const Array<OneD, NekDouble> &pInput,
-                             Array<OneD, NekDouble> &pOutput, const bool &flag)
-{
-    boost::ignore_unused(flag);
-    Vmath::Vcopy(pInput.size(), pInput, 1, pOutput, 1);
-}
-
-/**
- *
- */
-void PreconCfs::DoPreconCfs(
-    const Array<OneD, MultiRegions::ExpListSharedPtr> &pFields,
-    const Array<OneD, NekDouble> &pInput, Array<OneD, NekDouble> &pOutput,
-    const bool &flag)
-{
-    ASSERTL0(pInput.size() == pOutput.size(),
-             "In and Out not the same size in DoNullPrecon");
-    v_DoPreconCfs(pFields, pInput, pOutput, flag);
-    m_PreconTimesCounter++;
-}
-
-void PreconCfs::v_DoPreconCfs(
-    const Array<OneD, MultiRegions::ExpListSharedPtr> &pFields,
-    const Array<OneD, NekDouble> &pInput, Array<OneD, NekDouble> &pOutput,
-    const bool &flag)
-{
-    boost::ignore_unused(pFields, pInput, pOutput, flag);
-    NEKERROR(ErrorUtil::efatal, "v_DoPreconCfs not defined");
-}
-
-void PreconCfs::v_BuildPreconCfs(
-    const Array<OneD, MultiRegions::ExpListSharedPtr> &pFields,
-    const Array<OneD, const Array<OneD, NekDouble>> &intmp,
-    const NekDouble time, const NekDouble lambda)
-{
-    boost::ignore_unused(pFields, intmp, time, lambda);
-    NEKERROR(ErrorUtil::efatal, "v_BuildPreconCfs not defined");
-}
-
-bool PreconCfs::v_UpdatePreconMatCheck(const Array<OneD, const NekDouble> &res,
-                                       const NekDouble dtLambda)
-{
-    boost::ignore_unused(res, dtLambda);
-    NEKERROR(ErrorUtil::efatal, "v_UpdatePreconMatCheck not defined");
-    return false;
 }
 
 } // namespace Nektar